refactor(auth)!: de-bless 23 internal notebooklm.auth.* re-exports (#1592)
PR-1 of the auth public-surface tightening (sibling to the merged rpc tranche
#1589/#1591). Shrinks `auth.__all__` 46 → 23 by de-blessing the 23 names that
only first-party src/tests imported (never cli/_app) — migration leftovers from
the `_auth/*` extraction (ADR-0003 → ADR-0014).

Mechanism (same as the rpc tranche): names are removed from `__all__` but kept
importable as `notebooklm.auth.<name>` (the facade re-exports by assignment) for
back-compat; first-party src/tests now import them from their canonical
`notebooklm._auth.<sub>` home; one `removed-export` allowance each (baseline
v0.7.0, all 23 present there; `--check-stale` guards regrowth + forces prune at
the next release).

- Flip 8 core-src import sites (7 files) + the 23 names across 14 test files to
  `_auth.<sub>` (core/runtime may import `_auth`; cli/_app boundary untouched).
- Shrink `auth.__all__` + `EXPECTED_AUTH_ALL` 46 → 23; fix the `auth.py`
  docstring (`load_httpx_cookies` example) and the now-stale re-export comments.
- Remove the 22 de-blessed names present in the move-safety manifest
  (`_AUTH_FIRST_PARTY_COMPATIBILITY_NAMES`) — a deliberate change its docstring
  sanctions; add `test_auth_deblessed_names_stay_importable_but_unblessed`
  freezing that the 23 stay importable but absent from `__all__`.
- 23 `removed-export` allowlist entries; CHANGELOG `### Changed` + deprecations
  note.

Documented imports (`AuthTokens`, `convert_rookiepy_cookies_to_storage_state`,
the cookie-domain constants) and cohesive ops (`Account`, `enumerate_accounts`,
`fetch_tokens_with_domains`, `fetch_tokens_passive`, …) are unchanged. No runtime
behavior change; mypy unaffected (implicit re-export).

The ~14 names forced public by cli/_app (boundary-forbidden from `_auth`) stay
public; a deeper service-interface refactor to hide them was evaluated and
deferred (limited encapsulation payoff while names stay importable — momus+oracle
review, claude+codex, two rounds).

Verified: audit --check-stale 0/0, ruff, mypy, full pytest green (only the
pre-existing untracked-file guardrail failures remain).
