+++ tests/unit/test_cookie_persistence.py
@@ -11,13 +11,15 @@ import pytest
 
 import notebooklm._cookie_persistence as persistence_module
 import notebooklm._runtime.lifecycle as lifecycle_module
-from notebooklm._cookie_persistence import CookiePersistence
-from notebooklm.auth import (
-    AuthTokens,
+from notebooklm._auth.storage import (
     CookieSaveResult,
     CookieSnapshot,
     CookieSnapshotKey,
     snapshot_cookie_jar,
+)
+from notebooklm._cookie_persistence import CookiePersistence
+from notebooklm.auth import (
+    AuthTokens,
 )
 from tests._helpers.client_factory import build_client_shell_for_tests
